solution reorganization
